requirements.txt
beautifulsoup4==4.11.2
requests==2.28.2
serpapi==0.1.5
python-dotenv==0.21.1
groq
pytest==7.2.2
asyncio
aiohttp
chardet

